[BUGFIX] Delegate for TransientBackend
In order to prevent non-string values from being serialized
in object-capable frontends, the MultilevelCacheBackend
is marked as a Transient backend which means it receives
raw values for caching. The delegating set methods then
check each backend if *it* implements TransientInterface
and provides either the raw value or serialized as appropriate.